+++ webapp/src/core/getApiBaseUrl.ts
@@ -1,3 +1,0 @@
-export default function (): string {
-  return `${window.location.protocol}//${window.location.hostname}:42224/api`
-}