+++ webapp/src/pages/Events/index.tsx
@@ -1,4 +1,0 @@
-import React from 'react'
-import EventStream from '../../components/EventStream'
-
-export default () => <EventStream />